requirements.txt
google-api-python-client==1.7.2
google-auth==1.8.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
numpy<1.24
sounddevice
torchcrepe
librosa
faiss-cpu
pyworld
praat-parselmouth
ffmpeg-python
resampy
ffmpeg_downloader
pathvalidate
audiolm_pytorch
fairseq
gradio
soundfile
devtools
rich_argparse
transformers
scipy
funcy
encodec
rich
chat_downloader
nltk
demoji
sentencepiece
ninja
safetensors